+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+
+### Added
+- DSL support for `EXISTS` / `NOT EXISTS` and `IN (subquery)` / `NOT IN (subquery)` with association-based or entity-based correlation (#23)
+
 ## [0.1.0-rc.1] - 2026-04-06
 
 ### Miscellaneous